first complete run fixes and adjustments, please check the following comment

+++ src/clustering/measuresAndVisualizations/visualizationAndPlot.py
@@ -31,6 +31,7 @@ def dendogram(df, listOfClusteringColumns):
         for clusteringColumn in clusteringType:
             if "hierarchical" in clusteringColumn:
                     hierarchicalType = clusteringColumn.split (" ")[0] 
+                    threshold = int(''.join(filter(str.isdigit, clusteringColumn.split (" ")[1]))) 
                     imageName = hierarchicalType + " " + clusteringColumn + " dendogram.png"
                     imgPath = path.join(getPlotsFolderPath(), "dendogram", imageName)
                     if not visualizazionAlreadyExists (imgPath):
@@ -38,10 +39,10 @@ def dendogram(df, listOfClusteringColumns):
                         plt.title(f"Dendrogram for Clustering Type: {hierarchicalType} {clusteringColumn}")
                         
                         linkage_matrix = getMidRunObject ("(just object) " + hierarchicalType + "LinkageMatrix")
-                        dendrogram(linkage_matrix, orientation="top")
+                        dendrogram(linkage_matrix, orientation="top", truncate_mode='lastp', p=threshold)
 
-                         # Hide x-axis labels
-                        plt.tick_params(axis='x', which='both', bottom=False, top=False, labelbottom=False)
+                        # Hide x-axis labels
+                        #plt.tick_params(axis='x', which='both', bottom=False, top=False, labelbottom=False)
                         
                         checkSubFoldersExists(imgPath)
                         plt.savefig(imgPath)
@@ -49,7 +50,7 @@ def dendogram(df, listOfClusteringColumns):
 
 def correlationMatrix(df, listOfClusteringColumns):
     #@RafaelUrbina
-    for clusteringType in listOfClusteringColumns:#[["", ""],["", ""]]
+    """for clusteringType in listOfClusteringColumns:#[["", ""],["", ""]]
         for clusteringColumn in clusteringType:#["", ""]
             if "hierarchical" in clusteringColumn: #clusteringColumn example: "hierarchicalGroupAverage threshold3 criterion=blabla"
                 hierarchicalType = clusteringColumn.split (" ")[0] #get just the clustering type since the formatting of the strings for hierarchical is hierarchicalType + ' ' + 'threshold' + n + ' ' + 'criterion=' + n
@@ -79,7 +80,7 @@ def correlationMatrix(df, listOfClusteringColumns):
                     plt.savefig(imgPath, dpi=100)
                     plt.close()
 
-            elif "mixtureGuassian" in clusteringColumn:
+            elif "mixtureGaussian" in clusteringColumn:
                 tempDfScal = copyAndScaleDataset (df, continuousFeatures)
                 gmmtype = clusteringColumn
                 imageName = gmmtype + " correlationMatrix.png"
@@ -102,41 +103,43 @@ def correlationMatrix(df, listOfClusteringColumns):
                     plt.savefig(imgPath, dpi=100)
                     plt.close()
             else:
-                pass
+                pass"""
              
-def clusterBarChart(df, listOfClusteringColumns, featuresToPlotOn):
+def clusterBarChart(df, listOfClusteringTypes, featuresToPlotOn):
+    return
     #@AlessandroCarella
-    for clusteringType in listOfClusteringColumns:
+    for clusteringType in listOfClusteringTypes:
         for clusteringColumn in clusteringType:
             if clusteringColumn in df.columns:
                 for categoricalColumn in featuresToPlotOn:
-                    # Group the data by the categorical column and the cluster column
-                    grouped = df.groupby([categoricalColumn, clusteringColumn]).size().unstack().fillna(0)
-
-                    # Create the plot
-                    fig, ax = plt.subplots()
-
-                    # Plot the data
-                    for cluster in grouped.columns:
-                        ax.bar(grouped.index, grouped[cluster], label=f'Cluster {cluster}')
-
-                    # Set plot labels and legend
-                    ax.set_xlabel(categoricalColumn)
-                    ax.set_ylabel('Count')
-                    ax.set_title(f'{categoricalColumn} for {clusteringColumn}')
-                    ax.legend()
-                    
-                    # Set the figure size to 21:9 (3440x1440 pixels)
-                    fig.set_size_inches(21, 9)
-
                     imgPath = path.join(getPlotsFolderPath(), "clusterBarChart", f'{categoricalColumn} for {clusteringColumn}.png')
-                    checkSubFoldersExists(imgPath)
-                    plt.savefig(imgPath, dpi=100)
-                    plt.close()
+                    if not visualizazionAlreadyExists (imgPath):
+                        # Group the data by the categorical column and the cluster column
+                        grouped = df.groupby([categoricalColumn, clusteringColumn]).size().unstack().fillna(0)
+
+                        # Create the plot
+                        fig, ax = plt.subplots()
+
+                        # Plot the data
+                        for cluster in grouped.columns:
+                            ax.bar(grouped.index, grouped[cluster], label=f'Cluster {cluster}')
+
+                        # Set plot labels and legend
+                        ax.set_xlabel(categoricalColumn)
+                        ax.set_ylabel('Count')
+                        ax.set_title(f'{categoricalColumn} for {clusteringColumn}')
+                        ax.legend()
+                        
+                        # Set the figure size to 21:9 (3440x1440 pixels)
+                        fig.set_size_inches(21, 9)
+
+                        checkSubFoldersExists(imgPath)
+                        plt.savefig(imgPath, dpi=100)
+                        plt.close()
 
 def similarityMatrix(df, listOfClusteringColumns):
     #@RafaelUrbina
-    for clusteringType in listOfClusteringColumns:#[["", ""],["", ""]]
+    """for clusteringType in listOfClusteringColumns:#[["", ""],["", ""]]
         for clusteringColumn in clusteringType:#["", ""]
             if "hierarchical" in clusteringColumn: #clusteringColumn example: "hierarchicalGroupAverage threshold3 criterion=blabla"
                 hierarchicalType = clusteringColumn.split (" ")[0] #get just the clustering type since the formatting of the strings for hierarchical is hierarchicalType + ' ' + 'threshold' + n + ' ' + 'criterion=' + n
@@ -227,5 +230,6 @@ def similarityMatrix(df, listOfClusteringColumns):
                     checkSubFoldersExists(imgPath)
                     plt.savefig(imgPath, dpi=100)
                     plt.close()
+                    
             else:
-                pass
+                pass"""
